\docType{data}
\name{mutatedReads}
\alias{mutatedReads}
\title{Mock Set of Mutated Reads}
\description{
The mock set of mutated reads consists of DNA barcodes from the
\code{supplierSet} dataset concatenated with random DNA bases. Each read has
been mutated to a certain degree (i.e., single bases have been substituted).

The set is only intended for the man page examples.
}
\format{A character vector of 10,000 DNA sequences.}
\keyword{datasets}

