\docType{data}
\name{supplierSet}
\alias{supplierSet}
\title{Mock Set of DNA Barcodes}
\description{
The mock set of DNA barcodes represents a set that could come from any sample
tag or index supplier. The set has a minimum Hamming distance of 3 and
therefore allows the correction of a single substitution.

The set is only intended for the manual examples and should not be used in real experiments.

}
\format{A character vector of 48 DNA barcodes.}
\keyword{datasets}

